refactor(backend-auth): extract auth-related logic into dedicated authentication app on the backend
