+++ backend/authentication/serializers.py
@@ -1,0 +1,43 @@
+from django.contrib.auth.models import User
+from rest_framework import serializers
+
+from photo_gallery.serializers import LikeSerializer
+
+
+class UserSerializer(serializers.ModelSerializer):
+    likes = LikeSerializer(many=True, read_only=True)
+
+    class Meta:
+        model = User
+        fields = ["id", "username", "email", "first_name", "last_name", "likes"]
+
+
+class RegisterSerializer(serializers.ModelSerializer):
+    """Serializer for user registration."""
+
+    password = serializers.CharField(write_only=True, min_length=8)
+    email = serializers.EmailField(required=True)
+    first_name = serializers.CharField(required=True, min_length=2, max_length=50)
+    last_name = serializers.CharField(required=True, min_length=2, max_length=50)
+
+    class Meta:
+        model = User
+        fields = [
+            "username",
+            "email",
+            "first_name",
+            "last_name",
+            "password",
+        ]
+
+    def validate_email(self, value):
+        if User.objects.filter(email=value).exists():
+            raise serializers.ValidationError("A user with this email already exists.")
+        return value
+
+    def create(self, validated_data):
+        password = validated_data.pop("password")
+        user = User(**validated_data)
+        user.set_password(password)
+        user.save()
+        return user
