+++ backend/photo_gallery/serializers.py
@@ -1,4 +1,3 @@
-from django.contrib.auth.models import User
 from rest_framework import serializers
 
 from .models import Like, Photo, Photographer
@@ -60,39 +59,3 @@ class LikeSerializer(serializers.ModelSerializer):
     class Meta:
         model = Like
         fields = ["id", "user", "photo", "photo_id", "created_at"]
-
-
-class UserSerializer(serializers.ModelSerializer):
-    likes = LikeSerializer(many=True, read_only=True)
-
-    class Meta:
-        model = User
-        fields = ["id", "username", "email", "first_name", "last_name", "likes"]
-
-
-class RegisterSerializer(serializers.ModelSerializer):
-    """Serializer for user registration."""
-
-    password = serializers.CharField(write_only=True, min_length=8)
-
-    class Meta:
-        model = User
-        fields = [
-            "username",
-            "email",
-            "first_name",
-            "last_name",
-            "password",
-        ]
-
-    def validate_email(self, value):
-        if User.objects.filter(email=value).exists():
-            raise serializers.ValidationError("A user with this email already exists.")
-        return value
-
-    def create(self, validated_data):
-        password = validated_data.pop("password")
-        user = User(**validated_data)
-        user.set_password(password)
-        user.save()
-        return user
